requirements.txt
tensorflow==2.19.0
opencv-python
numpy
matplotlib
pillow
tqdm
